Refactor: Organize ViewModel files
This commit moves `SearchListViewModel` and `SearchViewModel` to a new `viewModels` package within the `presentation.component` directory.

This change improves the project structure by grouping related ViewModel files, making them easier to locate and manage.

The following files were moved:
- `presentation/src/main/java/com/example/presentation/component/screens/searchscreen/SearchListViewModel.kt` -> `presentation/src/main/java/com/example/presentation/component/viewModels/SearchListViewModel.kt`
- `presentation/src/main/java/com/example/presentation/component/screens/SearchScreen/SearchViewModel.kt` -> `presentation/src/main/java/com/example/presentation/component/viewModels/SearchViewModel.kt`

Imports in related files were updated to reflect the new package structure.
