+++ presentation/src/main/java/com/example/presentation/worker/RecentSearchSyncWorker.kt
@@ -3,7 +3,7 @@ package com.example.presentation.worker
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.example.presentation.component.searchListUi.SearchListViewModel
+import com.example.presentation.component.viewModels.SearchListViewModel
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
